% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscar.R
\name{k_h0}
\alias{k_h0}
\title{Calculate K for H0}
\usage{
k_h0(r_al, r_seen, recall_target)
}
\arguments{
\item{r_al}{The number of relevant documents seen before drawing from the urn began}

\item{r_seen}{The total number of relevant documents seen.}

\item{recall_target}{The recall target.}
}
\value{
k The smallest integer compatible with our null hypothesis
}
\description{
`k_h0` calculates the smallest number of relevant documents there would
have to be in our urn, for us to have missed a given recall target
}
